Refactor remove by reply id
