+++ lib/collections/comments.js
@@ -4,6 +4,7 @@ import linkifyStr from 'linkifyjs/string'
 import timeTickService from '../services/time-tick'
 import userService from '../services/user'
 import comment from '../services/comment'
+import { addCollectionMethods } from './CommentCollectionMethods'
 
 export const CommentsCollection = new Mongo.Collection('comments')
 
@@ -195,20 +196,7 @@ CommentsCollection._calculateAverageRating = calculateAverageRating
 
 const getCount = (scope, field) => scope[field] && scope[field].length ? scope[field].length : 0
 
-// Public API
-CommentsCollection.findPublic = (selector = {}, userId, opts = {}) => {
-  if (!opts.sort) opts.sort = { createdAt: -1 }
-
-  return CommentsCollection.find({
-    ...selector,
-    ...comment.commentStatusSelector(userId)
-  }, opts)
-}
-
-CommentsCollection.findOnePublic = (_id, userId) => CommentsCollection.findOne({
-  _id,
-  ...comment.commentStatusSelector(userId)
-})
+addCollectionMethods(CommentsCollection)
 
 CommentsCollection.helpers({
   likesCount: function () {
