+++ lib/collections/methods/comments.js
@@ -4,18 +4,14 @@ import mediaService from '../../services/media'
 import userService from '../../services/user'
 import commentService from '../../services/comment'
 import { CommentsCollection } from '../comments'
+import { noOptOptions } from '../../noOptOptions'
 import {
   getMongoReplyFieldDescriptor,
   removeReplyByReplyId,
   adjustReplyByReplyId,
 } from '../../services/reply'
 
-export const noOptOptions = {
-  validate: false,
-  filter: false,
-  getAutoValues: false,
-  removeEmptyStrings: false
-}
+export { noOptOptions }
 
 /**
  * Call a meteor method with anonymous user id if there is as the last argument.
@@ -451,11 +447,10 @@ Meteor.methods({
       return
     }
 
-    doc.replies = removeReplyByReplyId(doc.replies, replyId, reply => reply.userId === userId)
-
     const findSelector = { _id: documentId }
 
-    CommentsCollection.update(findSelector, { $set: { replies: doc.replies }  }, noOptOptions)
+    CommentsCollection.removeReplyByReplyId(replyId, reply => reply.userId === userId)
+
     triggerEvent('reply', 'remove', Object.assign({}, reply, findSelector, {
       rootUserId: doc.userId,
     }))
