+++ lib/noOptOptions.js
@@ -1,0 +1,6 @@
+export const noOptOptions = {
+  validate: false,
+  filter: false,
+  getAutoValues: false,
+  removeEmptyStrings: false
+}
